Refactored load map args
