+++ Runtime/Scripts/LoadSGBMapArgs.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d310832b6044faa47b9cd8a7a56a2898
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
